requirements.txt
Flask==3.0.3
gunicorn==23.0.0
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
transformers>=4.39.1